requirements.txt
pandas
python-docx
fastapi
uvicorn
nltk
elasticsearch
python-multipart
requests
starlette
scikit-learn